+++ test/frameSpec.js
@@ -20,7 +20,7 @@
 */
 
 const test = require('tape');
-const beamcoder = require('../index.js');
+const beamcoder = require('../ts/index.js');
 const util = require('util');
 
 test('Create a frame', t => {
