+++ tsconfig.json
@@ -97,5 +97,6 @@
     /* Completeness */
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
-  }
+  },
+  "files": [ "ts/beamstreams.ts", "ts/index.ts", "ts/install_ffmpeg.ts" ]
 }
